创建DMS模块

创建DMS模块

mdl_create(
  mdl_name = "mdlStarNight",
  mdl_dir = "/home/hulilei/mdls/",
  demo_dir = "/home/hulilei/mdls/mdlDemo/",
  demo_name = "mdlDemo",
  type = c("Com", "Pkg", "r", "Server", "UI")
)

Arguments

mdl_name

模块名称

mdl_dir

模块目录

demo_dir

模板目录

demo_name

模板名称

type

模板类型

Value

返回值

Examples

mdl_create()
#> Warning: incomplete final line found on '/home/hulilei/mdls/mdlStarNight/mdlStarNightCom/DESCRIPTION'
#> Warning: incomplete final line found on '/home/hulilei/mdls/mdlStarNight/mdlStarNightCom/mdlStarNightCom_ui.R'
#> Updating mdlStarNightUI documentation
#> Warning: Version of roxygen2 last used with this package is 7.1.1.  You only have version 7.1.0
#> Loading mdlStarNightUI
#> Writing NAMESPACE
#> Writing NAMESPACE
#>   ✔  checking for file ‘/home/hulilei/mdls/mdlStarNight/mdlStarNightUI/DESCRIPTION’ (428ms)
#>   ─  preparing ‘mdlStarNightUI’:
#>      checking DESCRIPTION meta-information ...  ✔  checking DESCRIPTION meta-information
#>   ─  checking for LF line-endings in source and make files and shell scripts
#>   ─  checking for empty or unneeded directories
#> ─  building ‘mdlStarNightUI_1.0.0.tar.gz’
#>      
#> Running /usr/local/lib/R/bin/R CMD INSTALL \
#>   /tmp/RtmpVZKvrM/mdlStarNightUI_1.0.0.tar.gz --install-tests 
#> * installing to library ‘/tmp/RtmpG56MkX/temp_libpath4d123d012eb2’
#> * installing *source* package ‘mdlStarNightUI’ ...
#> ** using staged installation
#> ** R
#> ** byte-compile and prepare package for lazy loading
#> Warning: replacing previous import ‘shiny::runExample’ by ‘shinyalert::runExample’ when loading ‘tsui’
#> Warning: replacing previous import ‘dplyr::where’ by ‘pryr::where’ when loading ‘tsdo’
#> [1] "load finished!"
#> Warning: replacing previous import ‘stringr::str_sort’ by ‘tsdo::str_sort’ when loading ‘tsda’
#> Warning: replacing previous import ‘stringr::str_equal’ by ‘tsdo::str_equal’ when loading ‘tsda’
#> Warning: replacing previous import ‘openxlsx::writeData’ by ‘tsda::writeData’ when loading ‘tsui’
#> Warning: replacing previous import ‘stringr::str_sort’ by ‘tsdo::str_sort’ when loading ‘tsui’
#> Warning: replacing previous import ‘stringr::str_equal’ by ‘tsdo::str_equal’ when loading ‘tsui’
#> ** help
#> *** installing help indices
#> ** building package indices
#> ** testing if installed package can be loaded from temporary location
#> Warning: replacing previous import ‘shiny::runExample’ by ‘shinyalert::runExample’ when loading ‘tsui’
#> Warning: replacing previous import ‘dplyr::where’ by ‘pryr::where’ when loading ‘tsdo’
#> [1] "load finished!"
#> Warning: replacing previous import ‘stringr::str_sort’ by ‘tsdo::str_sort’ when loading ‘tsda’
#> Warning: replacing previous import ‘stringr::str_equal’ by ‘tsdo::str_equal’ when loading ‘tsda’
#> Warning: replacing previous import ‘openxlsx::writeData’ by ‘tsda::writeData’ when loading ‘tsui’
#> Warning: replacing previous import ‘stringr::str_sort’ by ‘tsdo::str_sort’ when loading ‘tsui’
#> Warning: replacing previous import ‘stringr::str_equal’ by ‘tsdo::str_equal’ when loading ‘tsui’
#> ** testing if installed package can be loaded from final location
#> Warning: replacing previous import ‘shiny::runExample’ by ‘shinyalert::runExample’ when loading ‘tsui’
#> Warning: replacing previous import ‘dplyr::where’ by ‘pryr::where’ when loading ‘tsdo’
#> [1] "load finished!"
#> Warning: replacing previous import ‘stringr::str_sort’ by ‘tsdo::str_sort’ when loading ‘tsda’
#> Warning: replacing previous import ‘stringr::str_equal’ by ‘tsdo::str_equal’ when loading ‘tsda’
#> Warning: replacing previous import ‘openxlsx::writeData’ by ‘tsda::writeData’ when loading ‘tsui’
#> Warning: replacing previous import ‘stringr::str_sort’ by ‘tsdo::str_sort’ when loading ‘tsui’
#> Warning: replacing previous import ‘stringr::str_equal’ by ‘tsdo::str_equal’ when loading ‘tsui’
#> ** testing if installed package keeps a record of temporary installation path
#> * DONE (mdlStarNightUI)
#> 
#> Updating mdlStarNightServer documentation
#> Warning: Version of roxygen2 last used with this package is 7.1.1.  You only have version 7.1.0
#> Loading mdlStarNightServer
#> Writing NAMESPACE
#> Writing NAMESPACE
#>      checking for file ‘/home/hulilei/mdls/mdlStarNight/mdlStarNightServer/DESCRIPTION’ ...  ✔  checking for file ‘/home/hulilei/mdls/mdlStarNight/mdlStarNightServer/DESCRIPTION’
#>   ─  preparing ‘mdlStarNightServer’:
#>    checking DESCRIPTION meta-information ...  ✔  checking DESCRIPTION meta-information
#>   ─  checking for LF line-endings in source and make files and shell scripts
#>   ─  checking for empty or unneeded directories
#> ─  building ‘mdlStarNightServer_1.0.0.tar.gz’
#>      
#> Running /usr/local/lib/R/bin/R CMD INSTALL \
#>   /tmp/RtmpVZKvrM/mdlStarNightServer_1.0.0.tar.gz --install-tests 
#> * installing to library ‘/tmp/RtmpG56MkX/temp_libpath4d123d012eb2’
#> * installing *source* package ‘mdlStarNightServer’ ...
#> ** using staged installation
#> ** R
#> ** byte-compile and prepare package for lazy loading
#> ** help
#> *** installing help indices
#> ** building package indices
#> ** testing if installed package can be loaded from temporary location
#> ** testing if installed package can be loaded from final location
#> ** testing if installed package keeps a record of temporary installation path
#> * DONE (mdlStarNightServer)
#> 
#> Updating mdlStarNightPkg documentation
#> Loading mdlStarNightPkg
#> Writing NAMESPACE
#> Writing NAMESPACE
#>      checking for file ‘/home/hulilei/mdls/mdlStarNight/mdlStarNightPkg/DESCRIPTION’ ...  ✔  checking for file ‘/home/hulilei/mdls/mdlStarNight/mdlStarNightPkg/DESCRIPTION’
#>   ─  preparing ‘mdlStarNightPkg’:
#>      checking DESCRIPTION meta-information ...  ✔  checking DESCRIPTION meta-information
#>   ─  checking for LF line-endings in source and make files and shell scripts
#>   ─  checking for empty or unneeded directories
#> ─  building ‘mdlStarNightPkg_1.0.0.tar.gz’
#>      
#> Running /usr/local/lib/R/bin/R CMD INSTALL \
#>   /tmp/RtmpVZKvrM/mdlStarNightPkg_1.0.0.tar.gz --install-tests 
#> * installing to library ‘/tmp/RtmpG56MkX/temp_libpath4d123d012eb2’
#> * installing *source* package ‘mdlStarNightPkg’ ...
#> ** using staged installation
#> ** R
#> ** byte-compile and prepare package for lazy loading
#> ** help
#> *** installing help indices
#> ** building package indices
#> ** testing if installed package can be loaded from temporary location
#> ** testing if installed package can be loaded from final location
#> ** testing if installed package keeps a record of temporary installation path
#> * DONE (mdlStarNightPkg)
#> 
#> Updating mdlStarNightr documentation
#> Loading mdlStarNightr
#> Writing NAMESPACE
#> Writing NAMESPACE
#>      checking for file ‘/home/hulilei/mdls/mdlStarNight/mdlStarNightr/DESCRIPTION’ ...  ✔  checking for file ‘/home/hulilei/mdls/mdlStarNight/mdlStarNightr/DESCRIPTION’
#>   ─  preparing ‘mdlStarNightr’:
#>    checking DESCRIPTION meta-information ...  ✔  checking DESCRIPTION meta-information
#>   ─  checking for LF line-endings in source and make files and shell scripts
#>   ─  checking for empty or unneeded directories
#>   ─  building ‘mdlStarNightr_1.0.0.tar.gz’
#>      
#> Running /usr/local/lib/R/bin/R CMD INSTALL \
#>   /tmp/RtmpVZKvrM/mdlStarNightr_1.0.0.tar.gz --install-tests 
#> * installing to library ‘/tmp/RtmpG56MkX/temp_libpath4d123d012eb2’
#> * installing *source* package ‘mdlStarNightr’ ...
#> ** using staged installation
#> ** R
#> ** tests
#> ** byte-compile and prepare package for lazy loading
#> ** help
#> *** installing help indices
#> ** building package indices
#> ** testing if installed package can be loaded from temporary location
#> ** testing if installed package can be loaded from final location
#> ** testing if installed package keeps a record of temporary installation path
#> * DONE (mdlStarNightr)
#> 
#> 
#> Attaching package: ‘mdlStarNightr’
#> The following object is masked from ‘package:mdlStarNightPkg’:
#> 
#>     add
#> [1] TRUE